+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+1.2.2
+- i18nService.translate parameters are properly escaped
+- angular version updated
+
+1.2.1
+- adding named params using dictionary
+
 1.2.0
 - added support for named parameters